# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FusionPBX API",
        "contact": {
            "email": "vongruz@protonmail.com"
        },
        "version": "0.0.1"
    },
    "paths": {
...
        "/extensions": {
            "get": {
                "tags": [
                    "Extension"
                ],
                "summary": "Get all extensions in domain",
                "description": "`TODO`, describe in docs and return only some fields available for other users,\nadd parameters in query to select contact info, extension",
                "operationId": "App_Http_Controllers_Api_ExtensionController::getAll",
                "parameters": [
                    {
                        "name": "includes[]",
                        "in": "query",
                        "description": "Relations to be attached",
                        "allowReserved": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "users"
                                ]
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`TODO Stub` Could not created domain",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ExtensionWithRelatedUsersSchema"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{extension_uuid}": {
            "get": {
                "tags": [
                    "Extension"
                ],
                "summary": "Get extension by ID",
                "operationId": "App_Http_Controllers_Api_ExtensionController::getById",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/extension_uuid"
                    },
                    {
                        "name": "includes[]",
                        "in": "query",
                        "description": "Relations to be attached",
                        "allowReserved": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "users"
                                ]
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`TODO Stub`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExtensionWithRelatedUsersSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "`TODO Stub` Could not ..."
                    }
                }
            }
        },
        "/extension": {
            "post": {
                "tags": [
                    "Extension"
                ],
                "summary": "Extension create",
                "description": "Creates an extension and attaches it to a user (optionally)",
                "operationId": "App_Http_Controllers_Api_ExtensionController::create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/Extension"
                                    },
                                    {
                                        "properties": {
                                            "users": {
                                                "type": "array",
                                                "items": {
                                                    "allOf": [
                                                        {
                                                            "properties": {
                                                                "user_uuid": {
                                                                    "description": "User id's to attach extension to. If not passed, then extension is attached to the current user",
                                                                    "type": "string",
                                                                    "format": "uuid"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    ]
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            },
                            "example": {
                                "Create an exnetsion": [],
                                "Create an exnetsion basic example": {
                                    "summary": "`TODO example`",
                                    "value": {
                                        "code": 403,
                                        "message": "登录失败",
                                        "data": null
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Extension created response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Extension"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "`TODO Stub` Could not created domain",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "messages": [
                                        "Missing admin user",
                                        "No password for email"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/extension/{extension_uuid}": {
            "put": {
                "tags": [
                    "Extension"
                ],
                "summary": "Updates an extension",
                "operationId": "App_Http_Controllers_Api_ExtensionController::update",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/extension_uuid"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExtensionCreateSchema"
                            },
                            "example": {
                                "Create a user": [],
                                "Create a user basic example": {
                                    "summary": "`TODO example`",
                                    "value": {
                                        "code": 403,
                                        "message": "登录失败",
                                        "data": null
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`TODO Stub` Success ...",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Extension"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "`TODO Stub` Could not ..."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Extension"
                ],
                "summary": "Delets an extension",
                "operationId": "App_Http_Controllers_Api_ExtensionController::delete",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/extension_uuid"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`TODO Stub` Success ...",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "messages": [
                                        "`TODO` Describe response"
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "`TODO Stub` Could not ..."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Extension": {
                "properties": {
                    "extension_uuid": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "domain_uuid": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "extension": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string",
                        "writeOnly": true
                    },
                    "accountcode": {
                        "type": "string"
                    },
                    "effective_caller_id_name": {
                        "type": "string"
                    },
                    "effective_caller_id_number": {
                        "type": "string"
                    },
                    "outbound_caller_id_name": {
                        "type": "string"
                    },
                    "outbound_caller_id_number": {
                        "type": "string"
                    },
                    "emergency_caller_id_name": {
                        "type": "string"
                    },
                    "emergency_caller_id_number": {
                        "type": "string"
                    },
                    "directory_first_name": {
                        "type": "string"
                    },
                    "directory_last_name": {
                        "type": "string"
                    },
                    "directory_visible": {
                        "type": "string",
                        "default": "true"
                    },
                    "directory_exten_visible": {
                        "type": "string",
                        "default": "true"
                    },
                    "limit_max": {
                        "type": "string",
                        "default": 5
                    },
                    "limit_destination": {
                        "type": "string",
                        "default": "error/user_busy"
                    },
                    "missed_call_app": {
                        "type": "string"
                    },
                    "missed_call_data": {
                        "type": "string"
                    },
                    "user_context": {
                        "type": "string"
                    },
                    "toll_allow": {
                        "type": "string"
                    },
                    "call_timeout": {
                        "type": "number",
                        "default": 30
                    },
                    "call_group": {
                        "type": "string"
                    },
                    "call_screen_enabled": {
                        "type": "string",
                        "default": "false"
                    },
                    "user_record": {
                        "type": "string"
                    },
                    "hold_music": {
                        "type": "string",
                        "default": "local_stream://default"
                    },
                    "auth_acl": {
                        "type": "string"
                    },
                    "cidr": {
                        "type": "string"
                    },
                    "sip_force_contact": {
                        "type": "string"
                    },
                    "nibble_account": {
                        "type": "string",
                        "default": null
                    },
                    "sip_force_expires": {
                        "type": "number",
                        "default": null
                    },
                    "mwi_account": {
                        "type": "string",
                        "default": null
                    },
                    "sip_bypass_media": {
                        "type": "string"
                    },
                    "unique_id": {
                        "type": "number",
                        "default": null
                    },
                    "dial_string": {
                        "type": "string",
                        "default": null
                    },
                    "dial_user": {
                        "type": "string",
                        "default": null
                    },
                    "dial_domain": {
                        "type": "string",
                        "default": null
                    },
                    "do_not_disturb": {
                        "type": "string",
                        "default": null
                    },
                    "forward_all_destination": {
                        "type": "string",
                        "default": null
                    },
                    "forward_all_enabled": {
                        "type": "string",
                        "default": null
                    },
                    "forward_busy_destination": {
                        "type": "string",
                        "default": null
                    },
                    "forward_busy_enabled": {
                        "type": "string",
                        "default": null
                    },
                    "forward_no_answer_destination": {
                        "type": "string",
                        "default": null
                    },
                    "forward_no_answer_enabled": {
                        "type": "string",
                        "default": null
                    },
                    "forward_user_not_registered_destination": {
                        "type": "string",
                        "default": null
                    },
                    "forward_user_not_registered_enabled": {
                        "type": "string",
                        "default": null
                    },
                    "follow_me_uuid": {
                        "type": "string",
                        "format": "uuid",
                        "default": null
                    },
                    "forward_caller_id_uuid": {
                        "type": "string",
                        "format": "uuid",
                        "default": null
                    },
                    "follow_me_enabled": {
                        "type": "string"
                    },
                    "follow_me_destinations": {
                        "type": "string"
                    },
                    "enabled": {
                        "type": "string",
                        "default": "true"
                    },
                    "description": {
                        "type": "string"
                    },
                    "absolute_codec_string": {
                        "type": "string",
                        "default": null
                    },
                    "force_ping": {
                        "type": "string"
                    },
                    "max_registrations": {
                        "type": "string"
                    }
                }
            },
            "User": {
                "properties": {
                    "user_uuid": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "domain_uuid": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string",
                        "writeOnly": true
                    },
                    "user_email": {
                        "type": "string",
                        "writeOnly": true
                    },
                    "contact_uuid": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "api_key": {
                        "type": "string",
                        "readOnly": true
                    },
                    "user_enabled": {
                        "type": "string"
                    },
                    "add_user": {
                        "type": "string"
                    },
                    "add_date": {
                        "type": "string"
                    }
                }
            },
...
            "ExtensionCreateSchema": {
                "allOf": [
                    {
                        "properties": {
                            "reseller_reference_code": {
                                "description": "Reseller reference code`TODO properly save on user creation`",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/Extension"
                    }
                ]
            },
            "ExtensionWithRelatedUsersSchema": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Extension"
                    },
                    {
                        "properties": {
                            "users": {
                                "type": "array",
                                "items": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    ]
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "parameters": {
...
            "extension_uuid": {
                "name": "extension_uuid",
                "in": "path",
                "description": "Extension UUID",
                "required": true,
                "schema": {
                    "type": "string",
                    "format": "uuid",
                    "example": "541f8e60-5ae0-11eb-bb80-b31e63f668c8"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_auth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}